Refactor to put ctxmenu related functions in seperate file
